+++ mobile/alix/test/picker_controller_test.dart
@@ -2,6 +2,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:alix_mobile/picker/picker_controller.dart';
+import 'package:alix_mobile/profile.dart';
 import 'package:alix_mobile/picker/picker_models.dart';
 import 'package:alix_mobile/picker/picker_port.dart';